Tidy up the head material some more; in particular, vertically align the "PilRC vXXX" header with the authors even in non-IE browsers.
Also added lots of &nbsp; to the list of options in "Usage:", so that
we don't get "[-R <br> RESFILE]".  Remaining unsolved is the problem
that some browsers may break lines at the hyphens, producing "[- <br>
StripLoc]".  Not sure how to fix this, other than by reverting to
using <pre> -- shudder.