Warn if corresponding glyphs in different densities in a FONTFAMILY are not in proportion (e.g., exactly double the size for a double density font).
I went back and forth on whether this should be a warning or an error,
but in the end was not 100% sure that the test here was never a little
too Draconian, so relaxed it to a warning.

Also record the first family member's metrics rather than the last,
on the assumption that that one will be the single density member.

(And a typo fix.)